requirements.txt
django==2.2
djangorestframework==3.11.0
psycopg2==2.8.4
django-phonenumber-field==4.0.0
django-cors-headers==3.2.1
phonenumbers==8.11.5
djangorestframework-simplejwt==4.4.0